fix: claude 모듈 코드 리뷰 13개 수정 항목 적용
Critical:
- agent_runner.py: allowed_tools 폴백 None으로 수정 (허용/금지 뒤집힘 버그)

Short-term:
- agent_runner.py: asyncio.get_event_loop() → time.monotonic() 통일
- session.py: datetime.now() → datetime.now(timezone.utc) UTC 통일
- types.py: Callable[..., None] → 명시적 시그니처 부여

Mid-term:
- executor.py: 하위 호환 래퍼 7개 제거, _process_result에서 직접 호출
- executor.py/diagnostics.py: Path(__file__).parents[4] → SEOSOYOUNG_RUNTIME 환경변수
- security.py: 미사용 모듈 제거
- executor.py: session_runtime 메서드 직접 바인딩 제거
- intervention.py: run_in_new_loop 사용 이유 docstring 추가
- agent_runner.py: ClaudeAgentRunner 하위 호환 alias 제거
- session_context.py: float(ts) → Decimal 정밀도 개선
- executor.py: _execute_once 인라인 콜백 → 별도 메서드 분리
- agent_runner.py: _classify_process_error alias 제거

+++ src/seosoyoung/slackbot/claude/session_context.py
@@ -5,6 +5,7 @@
 """
 
 import logging
+from decimal import Decimal
 from typing import Optional, Protocol, runtime_checkable
 
 logger = logging.getLogger(__name__)
@@ -53,8 +54,8 @@ def build_initial_context(
         source_type = "thread"
         all_messages = list(slack_messages)
 
-    # 시간순 정렬
-    all_messages.sort(key=lambda m: float(m.get("ts", "0")))
+    # 시간순 정렬 (Decimal로 float 정밀도 손실 방지)
+    all_messages.sort(key=lambda m: Decimal(m.get("ts", "0")))
 
     # 최대 7개로 제한 (가장 최근)
     if len(all_messages) > MAX_INITIAL_MESSAGES:
@@ -105,15 +106,15 @@ def build_followup_context(
     pending = channel_store.load_pending(channel_id)
     all_messages = _merge_messages(judged, pending)
 
-    # last_seen_ts 이후 메시지만 필터링
-    cutoff = float(last_seen_ts)
-    unseen = [m for m in all_messages if float(m.get("ts", "0")) > cutoff]
+    # last_seen_ts 이후 메시지만 필터링 (Decimal로 float 정밀도 손실 방지)
+    cutoff = Decimal(last_seen_ts)
+    unseen = [m for m in all_messages if Decimal(m.get("ts", "0")) > cutoff]
 
     # linked 체인 정보: unseen 메시지가 참조하는 이전 메시지도 포함
     linked_ts_set = set()
     for msg in unseen:
         linked_ts = msg.get("linked_message_ts")
-        if linked_ts and float(linked_ts) <= cutoff:
+        if linked_ts and Decimal(linked_ts) <= cutoff:
             linked_ts_set.add(linked_ts)
 
     # linked 메시지 수집 (already seen이지만 참조 대상으로 포함)
@@ -125,7 +126,7 @@ def build_followup_context(
 
     # 병합 후 시간순 정렬
     combined = _merge_messages(linked_messages, unseen)
-    combined.sort(key=lambda m: float(m.get("ts", "0")))
+    combined.sort(key=lambda m: Decimal(m.get("ts", "0")))
 
     # 최대 제한
     if len(combined) > MAX_FOLLOWUP_MESSAGES:
